Moved the InstructionHandler to its own file.
